+++ flask_unchained/bundles/babel/__init__.py
@@ -3,10 +3,13 @@ import re
 
 from flask import Blueprint, current_app, g, request
 from flask_babelex import Domain, gettext as _gettext, ngettext as _ngettext
-from flask_unchained import Bundle, FlaskUnchained
+from flask_unchained import Bundle, FlaskUnchained, DEV, TEST
 from speaklater import make_lazy_string
 from typing import *
 
+from .config import (Config as BabelBundleConfig,
+                     DevConfig as BabelBundleDevConfig,
+                     ProdConfig as BabelBundleProdConfig)
 from .extensions import Babel, babel
 
 
@@ -34,12 +37,16 @@ class BabelBundle(Bundle):
     def before_init_app(self, app: FlaskUnchained):
         app.jinja_env.add_extension('jinja2.ext.i18n')
         babel.locale_selector_func = self.get_locale
-        if app.config.get('ENABLE_URL_LANG_CODE_PREFIX'):
+        if app.config.get('ENABLE_URL_LANG_CODE_PREFIX',
+                          BabelBundleConfig.ENABLE_URL_LANG_CODE_PREFIX):
             app.url_value_preprocessor(self.lang_code_url_value_preprocessor)
             app.url_defaults(self.set_url_defaults)
 
     def after_init_app(self, app: FlaskUnchained):
-        if not app.config.get('LAZY_TRANSLATIONS'):
+        if not app.config.get(
+                'LAZY_TRANSLATIONS',
+                BabelBundleDevConfig.LAZY_TRANSLATIONS if app.env in {DEV, TEST}
+                else BabelBundleProdConfig.LAZY_TRANSLATIONS):
             app.jinja_env.install_gettext_callables(gettext, ngettext, newstyle=True)
         else:
             app.jinja_env.install_gettext_callables(lazy_gettext, lazy_ngettext,
@@ -51,7 +58,7 @@ class BabelBundle(Bundle):
         return f'/<{self.language_code_key}>' + rule
 
     def register_blueprint(self, app: FlaskUnchained, blueprint: Blueprint, **options):
-        if app.config.get('ENABLE_URL_LANG_CODE_PREFIX'):
+        if app.config.ENABLE_URL_LANG_CODE_PREFIX:
             url_prefix = (options.get('url_prefix', (blueprint.url_prefix or ''))
                                  .rstrip('/'))
             options = dict(**options,
@@ -60,11 +67,11 @@ class BabelBundle(Bundle):
             app.register_blueprint(blueprint, **options)
 
     def add_url_rule(self, app: FlaskUnchained, rule: str, **kwargs):
-        if app.config.get('ENABLE_URL_LANG_CODE_PREFIX'):
+        if app.config.ENABLE_URL_LANG_CODE_PREFIX:
             app.add_url_rule(self.get_url_rule(rule), register_with_babel=False, **kwargs)
 
     def get_locale(self):
-        languages = current_app.config['LANGUAGES']
+        languages = current_app.config.LANGUAGES
         return g.get(self.language_code_key,
                      request.accept_languages.best_match(languages))