+++ flask_unchained/bundles/celery/extensions/celery.py
@@ -40,8 +40,8 @@ class Celery(BaseCelery):
     def init_app(self, app):
         self.app = app
         self.main = app.import_name
-        self.__autoset('broker_url', app.config.get('CELERY_BROKER_URL'))
-        self.__autoset('result_backend', app.config.get('CELERY_RESULT_BACKEND'))
+        self.__autoset('broker_url', app.config.CELERY_BROKER_URL)
+        self.__autoset('result_backend', app.config.CELERY_RESULT_BACKEND)
         self.config_from_object(app.config)
         self.autodiscover_tasks(lambda: [bundle.module_name
                                          for bundle in app.unchained.BUNDLES])