+++ flask_unchained/bundles/session/extensions/session.py
@@ -17,12 +17,12 @@ class Session(BaseSession):
         app.session_interface.serializer = dill
 
     def _get_interface(self, app):
-        if app.config['SESSION_TYPE'] == 'sqlalchemy':
+        if app.config.SESSION_TYPE == 'sqlalchemy':
             return SqlAlchemySessionInterface(
-                db=app.config['SESSION_SQLALCHEMY'],
-                table=app.config['SESSION_SQLALCHEMY_TABLE'],
-                key_prefix=app.config['SESSION_KEY_PREFIX'],
-                use_signer=app.config['SESSION_USE_SIGNER'],
-                permanent=app.config['SESSION_PERMANENT'],
-                model_class=app.config['SESSION_SQLALCHEMY_MODEL'])
+                db=app.config.SESSION_SQLALCHEMY,
+                table=app.config.SESSION_SQLALCHEMY_TABLE,
+                key_prefix=app.config.SESSION_KEY_PREFIX,
+                use_signer=app.config.SESSION_USE_SIGNER,
+                permanent=app.config.SESSION_PERMANENT,
+                model_class=app.config.SESSION_SQLALCHEMY_MODEL)
         return super()._get_interface(app)